+++ timezones/tz_utils.py
@@ -10,12 +10,6 @@ Example usage (get a fixed offset timezone)::
     print tz_utils.get_timezone('GMT +10:00')
 
 
-Example usage (guess timezone by IP, required geoip2!)::
-
-    tz_utils.GEOIP_DATA_LOCATION = '/usr/local/geo_ip/GeoIP2-City.mmdb'
-    assert tz_utils.guess_timezone_by_ip('201.246.115.62') == 'Chile/Continental'
-
-
 Example usage (format timezone by name)::
 
     print tz_utils.format_tz_by_name('Europe/Copenhagen')
@@ -33,6 +27,7 @@ Example usage (is a timezone valid?)::
 :copyright: 2012 by Amir Salihefendic ( http://amix.dk/ )
 :license: MIT
 """
+
 from __future__ import annotations
 
 from datetime import datetime, timedelta, tzinfo
@@ -41,57 +36,12 @@ import zoneinfo as zi
 
 from . import _defs
 
-try:
-    import geoip2.database as geoip2_db
-
-    HAS_GEOIP2 = True
-except ImportError:
-    HAS_GEOIP2 = False
-
 # --- Exports ----------------------------------------------
 __all__ = [
     "get_timezone",
     "is_valid_timezone",
-    "GEOIP_DATA_LOCATION",
-    "guess_timezone_by_ip",
     "format_tz_by_name",
 ]
-
-# --- Specifies the location of GeoIP GeoLiteCity.dat database ---
-GEOIP_DATA_LOCATION = None
-
-
-# --- Functions ----------------------------------------------
-def guess_timezone_by_ip(ip: str, only_name: bool = False):
-    """Given an `ip` with guess timezone using geoip2.
-    Returns a tuple of (tz_offets, tz_name, tz_formatted).
-    `None` is returned if it can't guess a timezone.
-
-    For this to work you need to set tz_utils.GEOIP_DATA_LOCATION
-    You can get this database from http://www.maxmind.com/app/geolitecity
-
-    Example usage::
-
-        from timezones import tz_utils
-
-        tz_utils.GEOIP_DATA_LOCATION = "/usr/local/geo_ip/GeoLiteCity.dat"
-        assert tz_utils.guess_timezone_by_ip("201.246.115.62") == "Chile/Continental"
-
-    """
-    geo_lib = _get_geoip_lib()
-    if geo_lib:
-        try:
-            record = geo_lib.city(ip)
-            if record:
-                location = record.location
-                if location and location.time_zone:
-                    if only_name:
-                        return location.time_zone
-                    else:
-                        return format_tz_by_name(location.time_zone)
-        except Exception:
-            record = None
-    return None
 
 
 def get_timezone(tzname: str) -> tzinfo | None:
@@ -154,23 +104,6 @@ def get_last_datetime_without_dst(tz: tzinfo):
 
 
 # --- Private ----------------------------------------------
-GEO_IP = None
-
-
-def _get_geoip_lib():
-    global GEO_IP
-
-    if not HAS_GEOIP2 or not GEOIP_DATA_LOCATION:
-        return None
-
-    try:
-        GEO_IP = geoip2_db.Reader(GEOIP_DATA_LOCATION)
-    except Exception:
-        return None
-
-    return GEO_IP
-
-
 _zero = timedelta(0)